refactor(identification): one module per scheme
Split the 1,704-line identification.py into impulso/identification/ —
the deletion test showed five independent adapters sharing a file. Code
moved verbatim; __init__ re-exports every name in use, so no importing
file changes. The accidental cross-class dependency dies: ProxySVAR and
ZeroSignRestriction no longer reach into SignRestriction for shock-coord
padding — _shared.pad_shock_coords owns it, and the two tests that
called the private staticmethod now exercise the public shock_coords
interface. _companion_spectral_radius lives with its sole consumer
(long_run); _PosteriorCache/_CACHE_MISS in _cache.
